# api/docs/swagger.yaml
basePath: /
definitions:
  handler.Credentials:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  model.APIKey:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      req_count:
        type: integer
      tier:
        type: string
      user_id:
        type: string
    type: object
  model.ArtistRef:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  model.CertificationFull:
    properties:
      artist:
        $ref: '#/definitions/model.ArtistRef'
      certified_at:
        type: string
      country:
        type: string
      created_at:
        type: string
      id:
        type: string
      level:
        type: string
      release:
        $ref: '#/definitions/model.ReleaseRef'
      sales_equivalent:
        type: integer
      source:
        type: string
    type: object
  model.ReleaseRef:
    properties:
      id:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  service.AuthResult:
    properties:
      api_key:
        $ref: '#/definitions/model.APIKey'
      plain_key:
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
host: localhost:8000
info:
  contact: {}
  description: Query music certifications (gold, platinum, diamond) from SNEP and
    other sources.
  title: Music Certifications API
  version: "1.0"
paths:
  /v1/artists/{id}/certifications:
    get:
      parameters:
      - description: Artist ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.CertificationFull'
                type: array
              meta:
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List certifications for an artist
      tags:
      - certifications
  /v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email and password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.AuthResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login and rotate API key
      tags:
      - auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email and password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.Credentials'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/service.AuthResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /v1/certifications:
    get:
      parameters:
      - description: Filter by country code (e.g. FR)
        in: query
        name: country
        type: string
      - description: Filter by level (e.g. platinum)
        in: query
        name: level
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.CertificationFull'
                type: array
              meta:
                type: object
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List certifications
      tags:
      - certifications
  /v1/certifications/{id}:
    get:
      parameters:
      - description: Certification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CertificationFull'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a certification by ID
      tags:
      - certifications
  /v1/me/apikey:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIKey'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get current user's API key info
      tags:
      - auth
  /v1/releases/{id}/certifications:
    get:
      parameters:
      - description: Release ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.CertificationFull'
                type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List certifications for a release
      tags:
      - certifications
  /v1/search:
    get:
      parameters:
      - description: Search query (artist or release name)
        in: query
        name: q
        required: true
        type: string
      - description: Filter by country code
        in: query
        name: country
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.CertificationFull'
                type: array
              meta:
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Search certifications
      tags:
      - certifications
  /v1/sources:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  type: string
                type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List available certification sources
      tags:
      - certifications
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"